Appending suggested changes to the pull request
I have removed the ^ and _ keyboard sequences which I had appended earlier, but now I have added a series with 𝐯 to incorporate the subscripted letters and Greek characters (lines 355-379). I have also taken the liberty to modify lines 347 and 348, to avoid any clash between ⇒ and 𝕝 (both formerly "dl"), and also between ⇐ and 𝕙 (both formerly "dh"). I have changed the combination for the italic 𝑛 to "In" to avoid clash with ∈, and the scripted 𝓃 to "Sn" the superscript ⁿ, and set the corresponding symbols ∈ and ⁿ to "in" and "sn" respectively. Finally, on lines 352 and 353, I have added two more symbols, which hope I have not added redundantly as before. These are the therefore and because (∴ and ∵) triple-dots. I ran searches for these unicodes inside my text editor and couldn't find them, so I considered it safe to add them. Do tell me if I have added what was already there in the file.